Fixed several error-check problems with bus config.  Changed names of a few Instrument internal members for clarity.  Added new Locked class and used it in one place in bus config code.  Fixed integer overflow problem with PField string conversion.  Stripped unused code out of inTraverse() and moved decls of variables for clarity.  Added beginnings of code needed to support multiple output devices.  Fixed compiler warning in table code.

+++ src/rtcmix/PField.cpp
@@ -37,7 +37,7 @@ PField::~PField()
 const char * 	
 PField::stringValue(double dindex) const
 {
-	const int ivalue = intValue(dindex);
+	const size_t ivalue = intValue(dindex);
 	return (const char *) ((ivalue < 0x100000) ? 0 : ivalue);
 }
